rename: [Step 5] Domain 테스트 코드 패키지 생성
