+++ src/test/java/racing/domain/CarTest.java
@@ -1,10 +1,9 @@
-package racing;
+package racing.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import racing.domain.Car;
 import racing.domain.strategy.MoveStrategy;
