+++ src/test/java/racing/domain/PositionTest.java
@@ -1,10 +1,9 @@
-package racing;
+package racing.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import racing.domain.Position;
 
 import static org.assertj.core.api.Assertions.*;
