+++ src/test/java/racing/domain/StringUtilTest.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.domain;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
