+++ src/test/java/racing/domain/TryCountTest.java
@@ -1,8 +1,7 @@
-package racing;
+package racing.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racing.domain.TryCount;
 
 import static org.assertj.core.api.Assertions.*;
